requirements.txt
asgiref==3.3.4
Django==3.2.2
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.14.0
django-cors-headers==3.14.0